Core/Battlefields: Fixed queue joining * Updated world handlers for battlefield Join.
ToDo
fix leave queue.
fix battlefields timers.
fix destructible game objects (wrong flags?)
Remove the wg group checks and set grouping methods...
THIER IS NO SERVER SET GROUPING FOR WORLDPVP!!!
that defeats the purpose, of having a zone at war.

+++ src/server/game/Server/WorldSession.h
@@ -797,15 +797,15 @@ class WorldSession
         void HandleInstanceLockResponse(WorldPacket& recvPacket);
 
         // Battlefield
-        void SendBfInvitePlayerToWar(uint32 BattleId, uint32 ZoneId, uint32 time);
+        void SendBfInvitePlayerToWar(uint32 BattleId, uint32 ZoneId, uint64 time);
         void SendBfInvitePlayerToQueue(uint32 BattleId);
         void SendBfQueueInviteResponce(uint32 BattleId, uint32 ZoneId);
         void SendBfEntered(uint32 BattleId);
         void SendBfLeaveMessage(uint32 BattleId);
-        void HandleBfQueueInviteResponse(WorldPacket &recv_data);
-        void HandleBfJoinQueue(WorldPacket &recv_data);
-        void HandleBfEntryInviteResponse(WorldPacket &recv_data);
-        void HandleBfExitRequest(WorldPacket &recv_data);
+        void HandleBfQueueInviteResponse(WorldPacket& recv_data);
+        void HandleBfJoinQueue(WorldPacket& recv_data);
+        void HandleBfEntryInviteResponse(WorldPacket& recv_data);
+        void HandleBfExitRequest(WorldPacket& recv_data);
 
         // Looking for Dungeon/Raid
         void HandleLfgSetCommentOpcode(WorldPacket& recv_data);
